backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-MySQLdb==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
PyMySQL==1.1.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
pandas==2.1.0
numpy==1.26.0
python-dateutil==2.8.2
Flask-SocketIO==5.3.4
eventlet==0.33.3
